Correct test

+++ moderngl_window/opengl/projection.py
@@ -70,7 +70,7 @@ class Projection3D:
         self._far = far or self._far
 
         self._matrix = glm.perspective(
-            self._fov, self._aspect_ratio, self._near, self._far, dtype="f4",
+            self._fov, self._aspect_ratio, self._near, self._far
         )
         self._matrix_bytes = self._matrix.to_bytes()
